refactor: change naming and delete button styles
