+++ mw-webapp/src/pages/authPage/AuthPage.module.scss
@@ -1,0 +1,7 @@
+.container {
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+}
