+++ mw-webapp/src/pages/authPage/AuthPage.tsx
@@ -1,21 +1,21 @@
 import {Link} from "react-router-dom";
-import styles from "src/logic/welcomePage/WelcomePage.module.scss";
+import styles from "src/pages/authPage/AuthPage.module.scss";
 
-export const WelcomePage = () => {
+export const AuthPage = () => {
   return (
     <div className={styles.container}>
       <Link to={"/"}>
-        <button className={styles.button}>
+        <button>
           Sign In
         </button>
       </Link>
       <Link to={"/"}>
-        <button className={styles.button}>
+        <button>
           Sign Up
         </button>
       </Link>
       <Link to={"main"}>
-        <button className={styles.button}>
+        <button>
           Workflow Page
         </button>
       </Link>
